Return EmptyBlockBuilder
