+++ shared_model/validators/block_validator.hpp
@@ -20,7 +20,7 @@
 
 #include "interfaces/iroha_internal/block.hpp"
 #include "validators/container_fields/container_validator.hpp"
-#include "validators/container_fields/transactions_validator.hpp"
+#include "validators/container_fields/non_empty_transactions_validator.hpp"
 
 namespace shared_model {
   namespace validation {
@@ -35,7 +35,7 @@ namespace shared_model {
         : public ContainerValidator<
               interface::Block,
               FieldValidator,
-              TransactionsValidator<TransactionValidator>> {
+              NonEmptyTransactionsValidator<TransactionValidator>> {
      public:
       /**
        * Applies validation on block
@@ -43,11 +43,11 @@ namespace shared_model {
        * @return Answer containing found error if any
        */
       Answer validate(const interface::Block &block) const {
-        return ContainerValidator<
-            interface::Block,
-            FieldValidator,
-            TransactionsValidator<TransactionValidator>>::validate(block,
-                                                                   "Block");
+        return ContainerValidator<interface::Block,
+                                  FieldValidator,
+                                  NonEmptyTransactionsValidator<
+                                      TransactionValidator>>::validate(block,
+                                                                       "Block");
       }
     };
