+++ test/module/shared_model/builders/protobuf/test_empty_block_builder.hpp
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "module/shared_model/builders/protobuf/empty_block_template.hpp"
+#include "builders/protobuf/builder_templates/empty_block_template.hpp"
 #include "module/shared_model/validators/validators.hpp"
 
 #ifndef IROHA_TEST_EMPTY_BLOCK_BUILDER_HPP
@@ -13,9 +13,9 @@
  * Builder alias, to build shared model proto empty block object avoiding
  * validation and "required fields" check
  */
-using TestEmptyBlockBuilder =
-    TemplateEmptyBlockBuilder<(1 << TemplateEmptyBlockBuilder<>::total) - 1,
-                              shared_model::validation::AlwaysValidValidator,
-                              shared_model::proto::EmptyBlock>;
+using TestEmptyBlockBuilder = shared_model::proto::TemplateEmptyBlockBuilder<
+    (1 << shared_model::proto::TemplateEmptyBlockBuilder<>::total) - 1,
+    shared_model::validation::AlwaysValidValidator,
+    shared_model::proto::EmptyBlock>;
 
 #endif  // IROHA_TEST_EMPTY_BLOCK_BUILDER_HPP
